+++ src/utils/getWrapperSourceLink.js
@@ -1,7 +1,7 @@
 import {
-  ADDRESSES_SCOPE,
   FIND_ALL_CQL,
   PER_REQUEST_LIMIT,
+  TENANT_ADDRESSES_API,
 } from './constants';
 
 export const getWrapperSourceLink = (key, resourceLimit = PER_REQUEST_LIMIT) => {
@@ -9,7 +9,7 @@ export const getWrapperSourceLink = (key, resourceLimit = PER_REQUEST_LIMIT) => 
     PREFIXES: `/orders/configuration/prefixes?limit=${resourceLimit}&query=${FIND_ALL_CQL} sortby name`,
     SUFFIXES: `/orders/configuration/suffixes?limit=${resourceLimit}&query=${FIND_ALL_CQL} sortby name`,
     ACQUISITIONS_UNITS: `/acquisitions-units/units?limit=${resourceLimit}&query=${FIND_ALL_CQL} sortby name`,
-    ADDRESSES: `/settings/entries?limit=${resourceLimit}&query=(scope==${ADDRESSES_SCOPE}) sortBy value`,
+    ADDRESSES: `/${TENANT_ADDRESSES_API}?sortBy name`,
     CONTRIBUTOR_NAME_TYPES: `/contributor-name-types?limit=${resourceLimit}&query=${FIND_ALL_CQL} sortby name`,
     IDENTIFIER_TYPES: `/identifier-types?limit=${resourceLimit}&query=${FIND_ALL_CQL} sortby name`,
     ACQUISITION_METHODS: `/orders/acquisition-methods?limit=${resourceLimit}&query=${FIND_ALL_CQL} sortby value`,